requirements.txt
amqp==2.2.2
billiard==3.5.0.3
celery==4.1.0
certifi==2018.1.18
chardet==3.0.4
Django==2.0.1
django-elasticsearch-dsl==0.4.4
djangorestframework==3.7.7
elasticsearch==5.5.2
elasticsearch-dsl==5.4.0
idna==2.6
ipaddress==1.0.19
kombu==4.1.0
oauthlib==2.0.6
python-dateutil==2.6.1
pytz==2017.3
requests==2.18.4
requests-oauthlib==0.8.0
six==1.11.0
tweepy==3.5.0
urllib3==1.22
vine==1.1.4
